Prevent Facebook login popup from being blocked
Facebook's login popup was being blocked because FB.login was being
called async from FB.getLoginStatus's callback. In order to get around
this, the user's login status is checked when the SDK is loaded and it
is saved for future reference.

Also, it is important to subscribe to the FB auth event rather than
doing a one-time check of the auth status because if the user logs out
of Facebook while using CA, it will result in an error upon querying the
graph API. If we're subscribed to the auth event, the error will
automatically update the fbAuth status which will prompt the user to log
in again.